requirements.txt
pydantic
opencv-python
pillow
xarm-python-sdk
google-genai
pykinect-azure
pygame_gui
sounddevice
scipy
pybullet